requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
gradio==4.44.0
pydantic==2.12.5
python-dotenv==1.2.1
openai==2.16.0
reportlab==4.4.9
httpx==0.28.1
